refactor: move helper function to /utils
